Latest

+++ workspace/linux-support/src/network_device/c/DOWNSHIFT_DEV_.rs
@@ -2,6 +2,6 @@
 // Copyright © 2020 The developers of linux-support. See the COPYRIGHT file in the top-level directory of this distribution and at https://raw.githubusercontent.com/lemonrock/linux-support/master/COPYRIGHT.
 
 
-pub(crate) const DOWNSHIFT_DEV_DEFAULT_COUNT: u32 = 0xFF;
+pub(crate) const DOWNSHIFT_DEV_DEFAULT_COUNT: u8 = 0xFF;
 
-pub(crate) const DOWNSHIFT_DEV_DISABLE: u32 = 0;
+pub(crate) const DOWNSHIFT_DEV_DISABLE: u8 = 0;
